CFD: Kill past connections from rudder-server
SQL statements in rudder-server are not executed with a timeout context,
instead we are letting them take as much time as they need.

Due to the above, when a server shutdown is initiated in a cloud environment
while long-running statements are being executed, the server process will not
manage to shutdown gracefully, since it will be blocked by the SQL statements.
The container orchestrator will eventually kill the server process, leaving
one or more dangling connections in the database.

This change ensures that before a new rudder-server instance starts working,
all previous dangling connections belonging to this server are being killed.

To support killing only connections opened by a specific rudder-server,
the application_name parameter is being populated with the server's hostname.
This behaviour can be overrided though through the JOBS_DB_APP_NAME environment
variable.