Fix: Jobsdb migration bug (#1785)
Fix: Jobsdb migration bug

If server crashes (panics)  while the migrateDSLoop is running and has managed
to create a new dataSet but has not managed to mark the start of copy operation
in the journal yet, then the dataSet's tables are not cleaned up during recovery.

This fix reverses the order of actions so that the copy operation is recorded
in the journal before any new dataSet gets created. To achieve this, HandleT#addNewDS
method had to be refactored to accept the dataSetT to be created as a parameter
instead of calculating it within the method's body  based on the previous method parameters.
Due to this change, all invocations of this method had to be adapted as well.

Hint:
The jobDB's journal table appears to be a poor man's method for performing transaction demarcation.
To understand this a bit better, please consider first that long-running transactions
are not preferred by rudder-server due to the involved lock contention.
But, there are occasions where we need to perform multiple database
operations as atomic operations (either succeed completely or fail completely). An operation entry
in the journal table marks the start of such a transaction. If the transaction fails to complete
successfully, the server will crash (panic). Upon restart, the recovery process will kick in, it
will scan the journal for incomplete operations and perform all necessary compensation actions, a.k.a. rollback.
After recovery, the server's services are safe to start again against a consistent state.