Fix: execute ANALYZE after completing batch jobsdb operations
During a migration or import, new tables get created and populated in bulk with a large amount of rows.
It is possible that the AUTOVACUUM process does not manage to analyze those tables before we start querying them
which will result in the database's query planner making wrong decisions that may lead to never-ending queries.

This fix makes sure that whenever data are inserted in bulk in JobsDB, affected tables are explicitly ANALYZEd
before we release them into the wild