Fix: EventBus memory issue with slow subscribers (#1802)
Whenever a new event arrives, the EventBus spawns a new goroutine for every subscriber, which tries to push the event to the subscriber's channel.
In case of a slow subscriber, i.e. a subscriber who consumes at a slower pace than new events arrive, multiple goroutines end up waiting to
publish overdue events which are kept in memory. This scenario happens frequently in Hosted environment, where configuration structures which
are being passed as events are large (60MB-100MB) and ultimately cause the process to be OOMKilled.

This fix changes the following:
- Don't send all configuration changes to slow subscribers, only send the most recent one and discard intermediate configurations.
- Treat topics as Behaviour Subjects (rx pattern). Each topic remembers its last value and whenever a new subscription arrives the latest value gets published automatically.