feat: calculate aggregated pending_events_count metrics for all workspaces
Aggregating metrics for all workspaces in InfluxDB through Grafana using a subquery is suboptimal and essentially prohibitive, performance-wise.
Setting up separate alerts per workspace is also not possible due to the high cardinality of workspaces in the hosted environment.
For all the above reasons, we need to calculate aggregated metrics along with the high-granularity ones.