feat(router): use a circuit breaker while producing custom destination clients (#2016)
In case of wrong configuration, clients are not being created successfully and their creation is being retried on every new event that arrives at the router.
Any delay in client creation causes cascading delays, since there can be multiple workers for the same destination, all of them waiting their turn to acquire a lock so that they can try to create the missing client introducing additional delays for the next workers in the queue.

For that reason we are introducing a circuit breaker in the client creation process. If the system fails to create a client for 6 consecutive times, it will stop trying (immediately returning an error) for the next 60 seconds. After this timeout, it will try once, if it fails again the circuit will reopen for another 60 seconds etc.