fix: aws session creation is failing for s3 manager when roles are used
Redshift to use awsSession.Config.Credentials.Get() when roles
are used to get temporary credentials instead of sts.GetSessionToken.

Unify AWS session config creation process so that it is consistent
across all AWS destinations. Earlier S3 has slightly different process
to create SessionConfing so when updated the session config creation in
aws utils to support RoleBasedAuth flag, it got missed for S3 manager
as it was using different implementation so this unification should
prevent such errors in the future.