fix: snowpipe backoff missing for validation error (#5504)
# Description

If a Snowpipe destination is created with an invalid private key, the
Snowpipe client returns an `ERR_VALIDATION_ERROR` code. Previously, this
was handled in the default case, preventing it from being classified as
an `Authz` error. As a result, neither backoff nor config validation
occurred in the Upload function.

This PR fixes the issue by correctly identifying and returning this
error as `Authz`.

- Added `ErrAuthenticationFailed` for invalid user name and
`ErrRoleDoesNotExistOrNotAuthorized` for invalid role

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.