feat: gRPC API to expire the warehouse schema of a destination (#5508)
# Description

This PR adds a GRPC API which expires all warehouse schemas of a
destination in the `wh_schemas` table by setting `expiresAt` column as
`now()`. This will ensure that the next time these schemas are fetched
from the DB, they will re-fetched from the warehouse.

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.