fix(jobsdb): clearDB option fails with error (#5570)
# Description

Lack of tests on the `WithClearDB` option caused this feature to break
after we introduced migrations that run always as well as due to other
prior code refactoring. This option is required for proper execution of
jobsdb benchmarks, thus fixing it and adding a test to handle future
regressions.

## Linear Ticket

resolves PIPE-1945

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.