refactor: add columns in snowflake using IF NOT EXISTS (#5581)
# Description

Currently, when adding columns in Snowflake, we use `ADD COLUMN`, which
results in an error if the column already exists. For single-column
additions, we catch this error.

This PR updates the query to use `IF NOT EXISTS`, preventing Snowflake
from throwing an error for existing columns and removing the need for
error handling.

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.