chore: increase max idle connections per host for kinesis (#5652)
# Description

Kinesis is causing high CPU usage due to high number of TLS handshakes
happening. Solution is to increase number of connections that can stay
idle leaving room for not killing the connections once done which will
reduce TLS connections.
To begin with we are increasing the number of connections from 2 to 64.

## Linear Ticket

Resolves INT-3397

https://linear.app/rudderstack/issue/INT-3397/evaluate-workarounds-for-high-tls-connections-in-expressvpn-due-to

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.