feat: use header augmenter for router transformer client (#5640)
# Description

We are introducing a header augmenter for router transformer clients for
OAuth destinations. It will improve performance by not iterating each
event and adding the secret to the body. It will add the secret to
header and from transformer we will spread to every body.

## Linear Ticket


https://linear.app/rudderstack/issue/INT-3390/consider-alternative-key-injection-method-in-payloads

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.

---------

Co-authored-by: Dilip Kola <kdilipkola@gmail.com>