fix: jobsdb payload limit used compressed column sizes instead of actual payload length (#5723)
Job queries aggregated over `pg_column_size(event_payload)` to limit the
result when a size limit is passed. This could lead to loading more jobs
into memory than can be handled safely because it's the compressed size
of the column(when compressed). Using the `octet_length(event_payload)`
instead to find the actual byte size of the event payload column.