chore: change jobsdb payload column type default to text (#5746)
Change jobsdb column type default to text. This change ensures - when
not configured new datasets created have a `text` event_payload column.
Left the config and related handling for now so that it can phased out
at a later point so it leaves some time for jobsdb to compact and
cleanup older jsonb column containing datasets.