fix: use copy to load table in append mode for redshift(#5743)
# Description
- `tx.Commit` for redshift getting stuck in some cases when loading data
into table. In case of append mode, we can directly load file instead of
doing it in transaction
- Since we are now skipping creating staging table if append mode is
preferred. In case of `loadUserTables`, it relies on indentifies staging
table because `users` table will be deduped/merged even when append mode
is preferred. So, creating a staging table for `identifies` is now
handled by `loadUserTables`

## Linear Ticket
Part of WAR-517, WAR-515

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.