chore(jobsdb): remove jobMinRowsMigrateThres configuration option (#5763)
# Description

- `jobMinRowsLeftMigrateThreshold` is now the only configuration option
that controls whether a table is small or not, i.e. the migration logic
will expect two small tables to exists for an actual migration to happen
(a.k.a ds compaction).
- Including equality check during `jobDoneMigrateThres` comparison, so
that it is easier to disable this option by setting it to 1. It is now
recommended to turn off `jobDoneMigrateThres`, so that datasets only get
migrated whenever a ds compaction is possible. Migrating an incomplete
dataset requires time and acquires a lock (stops-the-world), thus if
this is to lead to the same number of datasets, it doesn't provide much
benefit.

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.