fix(jobsdb): completed datasets don't get deleted without a pair (#5793)
# Description
Right now we are keeping one extra dataset alive even though it is
completed, since it needs a pair for a migration to happen.
A completed dataset should be eligible for migration and shouldn't
require a pair for it to be migrated (deleted)

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.