feat: support new consent resolution strategy values (#5798)
# Description

Besides, "and" and "or" consent resolution strategies, we're now
supporting the new "all" and "any" values. The UI changes will be rolled
out very soon.

Also, the logic now prefers to use "allowed consent IDs" and falls back
to "denied consent IDs" to align with the client side implementation.
It's a bit tricky with allowed consent IDs as some older SDK versions
send it as an object for OneTrust CMP.
So, we're handling both an object and array of strings.

Additionally, I've sanitized the consent IDs from the config and event
before resolving them for filtering events to destinations.


## Linear Ticket

https://linear.app/rudderstack/view/scrum-board-75b83c60706a5

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.