chore(jobsdb): remove support for jobDoneMigrateThres (#5873)
# Description

This configuration option is already disabled in production (by setting
it to 1). Removing it, since the cost of migrating pending jobs without
reducing the number of datasets is not worth taking.

**Coming next:** Avoid migrating a small and incomplete dataset (needing
a pair) if all other datasets to be migrated are complete (without
pending jobs). E.g. if dataset1 is complete and dataset2 needs a pair,
only dataset1 will be migrated/deleted. Dataset2 will only be migrated
if there is another dataset3 that needs a pair too.

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.