fix(processor): throughput stalling due to mutex deadlock on full connection pool (#5919)
# Description

When `Processor.enableConcurrentStore` is enabled, the processor
performs the following operations concurrently:

- Storing batchrouter jobs
- Storing router jobs
- Storing proc error jobs
- Updating gateway job statuses (transaction gets committed only after
all above transactions are successfully committed)

Each of these operations uses a separate database connection.

In the "store router jobs" operation, before opening a connection, a
mutex is acquired per destinationID to ensure that only one goroutine
inserts jobs for a given destination at a time.

However, this can lead to a deadlock under high load: if all database
connections in the pool are occupied by goroutines waiting to acquire a
destinationID mutex, and the one goroutine that has acquired the mutex
is blocked waiting for a database connection, progress stalls. The
result is a circular wait where no goroutine can proceed, causing the
processor's throughput to stall..

To fix this deadlock, whenever `Processor.enableConcurrentStore` is
enabled, we're acquiring the mutexes early, before any other connection.

Additionally, we're increasing the default connection pool size to 80
and limiting it to 20 for gateway.

## Linear Ticket

resolves PIPE-2101

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.