chore(processor): use pipeline-specific destination locks during store (#5992)
# Description

Include the pipeline index when we’re creating the key for
`storePlocker`. Since the same userID will always be processed by the
same processor pipeline index, we should be fine with this change and
wouldn't any cause out-of-order panics in router.

## Linear Ticket

resolves PIPE-2143

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.