fix: incorrect snowpipe channel id being returned if insert recreates the channel (#6012)
# Description

**BUG**
A fix was made for preventing duplicates in Snowpipe in PR #5957 . But
it missed returning the id of the newly created channel in the `Upload`
method. As a result, `Poll` method was still receiving the stale channel
id. So the GetStatus API call was failing with 404, causing the event to
be marked as failed. This failed event would be retried causing
duplicates

**FIX**
- The actual fix was simple to propagate the new channel id from the
Insert API method back to the caller. But there was another miss, that
post channel recreation the cache was not being updated. So instead this
PR, moves the recreation logic from the API layer to the manager layer.
This meant reverting the changes done in `api.go` and `apiadapter.go`
and the corresponding test files
- In case of 404 status code, the api insert method is now returning an
error `ErrChannelNotFound`
- Created a `insert` method in manager which has the channel recreation
logic based on the error type. This method is also being called from
`discards`
- Added a regression unit test