chore: use kitsync.ErrGroup instead of errgroup.Group (#6011)
# Description

Latest version of `errgroup.Group` introduces a [breaking
change](https://github.com/golang/go/issues/53757#issuecomment-2980442880),
which doesn't propagate a panic until `Wait` is called. Thus, wherever
we postpone the `Wait` call until shutdown, we are replacing usages of
`errgroup.Group` with `kitsync.ErrGroup` which is based on the previous
behaviour of the `errgroup.Group`.

We are not using `kitsync.EagerGroup` for the following reasons:
- A zero-value `kitsync.EagerGroup` is not valid
- It exhibits unique behaviour with respect to context cancellation
which can lead to propagating an error during `EagerGroup.Wait` even
though no goroutine returned an error. This has to do mostly with the
eager group avoiding to call goroutines if the context has been canceled
and it communicates this condition through an error.

## Linear Ticket

resolves PIPE-2154

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.