chore: bump go kit (#6034)
Description

With the v2 SDK, if the endpoint is passed as an empty string it is
using the empty endpoint to route the data. This is not going to affect
any customers but should rather affect tests. error index was one place
where we were passing an empty endpoint and this caused issues

## Linear Ticket

fixes PIPE-2167

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.