chore: drop bot events in processor after capturing metrics (#6052)
# Description

This PR moves bot event dropping from the ingestion service to the
processor to capture accurate reporting metrics before deduplication
occurs.

### Background
Previously, bot events were dropped at the ingestion service level. The
main issue was that there is currently no way to emit reporting metrics
from ingestion, so we were using Prometheus metrics to power the UI
which are not accurate. Now we are using reporting metrics in the
processor to capture accurate bot event statistics.


### Changes made
- Drop bot events in the processor when botAction is set to "drop" in
message parameters
- Block live events recording in gateway for bot events that will be
dropped in the processor
- Capture accurate bot event metrics before the deduplication stage

#### Additional changes
- Move event blocking logic to occur before dedup processing.
- Cleanup `botAction` empty check before enriching them.


## Linear Ticket

Resolves OBS-858

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.