fix: negative MTU count (#6180)
# Description

Fixes issue where MTU calculation could result in negative values due to
IDs being present in `idTypeIdentifiedAnonymousID` but not in
`idTypeUserID`.

## Problem

MTU is calculated using the formula: `cardinality of idTypeUserID -
cardinality of idTypeIdentifiedAnonymousID`

When some IDs are added to `idTypeIdentifiedAnonymousID` but not to
`idTypeUserID`, the cardinality of `idTypeIdentifiedAnonymousID` can
exceed that of `idTypeUserID`, resulting in negative MTU values.

## Solution

Ensure all IDs that are added to `idTypeIdentifiedAnonymousID` are also
added to `idTypeUserID`. This maintains the integrity of the MTU
calculation formula.

## Changes

- Fixed `generateReportsFromJobsLegacy` method to add `previousID` to
`idTypeUserID` HLL
- New changes are behind a config flag for controlled rollout
- Added ability to generate shadow metrics with different workspaceId
and sourceId for validation

More details in this
[doc](https://www.notion.so/rudderstacks/MTU-count-with-Alias-event-23bf2b415dd080b2babec029739eccfc?source=copy_link)

## Linear Ticket

Fixes OBS-876

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.