fix: overwrite payloadStage in error response for aborted jobs (#6193)
# Description

Currently, when a job is aborted, the payloadStage is incorrectly set to
"delivery" in error response. This PR overwrites the payloadStage in
such cases to reflect the correct stage i.e `router_input`.

## Linear Ticket

[Resolves
OBS-879](https://linear.app/rudderstack/issue/OBS-879/overwrite-payloadstage-in-sample-response-when-a-job-is-aborted)

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.