chore: throttler should emit no warning if not enabled (#6223)
# Description

Our only means of disabling throttling for a destination type or
destination id is by setting the limit or window to zero. Thus it makes
no sense to emit a warning log in case some throttler is disabled.
Renamed `validConfiguration` to `enabled` to avoid any confusion in the
future.

Regression introduced by #6181

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.