chore: stop collecting errors in proc_error jobsdb (#6232)
# Description

We are no longer be collecting errors in `proc_error` jobsdb as we are
planning to remove error collection for debugging purposes in the near
future.

What was happening until now:

- rudder-server was collecting failed jobs from webhook, processor
steps, router, and batchrouter, and was storing these failed jobs in the
`proc_error` JobsDB.
- periodically, the `stash` processor component was creating backup
files from the `proc_error` JobsDB and uploading them to
`rudder-proc-err-logs` in object storage.

In the past, we relied on `proc_error` jobs for replays, but that’s no
longer the case. At this point, it feels like we’re spending unnecessary
resources collecting and storing all this information which no-one
actually needs, using as an excuse that it is useful for debugging
purposes…

As a first step we are stopping collection and storage of failed jobs to
`proc_error`. In a follow up step we will be removing all relevant
`proc_error` code altogether.

## Linear Ticket

resolves PIPE-2295

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.