fix: ut mirroring tests (#6341)
# Description

We had data races in the UT mirroring tests because we were changing the
actual slice passed to the transformer instead of creating a copy.

Once that was fixed, the tests showed that we were using the wrong
golden file (i.e. `goldenUtMirrorClientEvents.json` should have had the
original client events and not the transformer response).

Additionally, if a difference was detected between the original
transformer and the mirrored one, we'd upload the original `eventList`
concurrently. To avoid further races we're not passing a hard copy.

## Linear Ticket

< [Linear
Ticket](https://linear.app/rudderstack/issue/PIPE-2385/ut-mirroring-test-failure)
>

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.