fix: use type field from payload to check if event is blocked (#6487)
# Description

Currently, we are using `msg.Properties.RequestType` as the event type
to check if an event is blocking. This is incorrect because RequestType
doesn’t represent the actual event type. This PR fixes it by relying on
the type field present in the payload instead.

## Linear Ticket

Fixes PIPE-2527

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.

---------