feat(jobsdb): introduce partition_id column (#6502)
# Description

JobsDB tables now have a new column, `partition_id`. 
This column will be indexed and populated in the following JobsDB
instances:
- `gw`
- `rt`
- `batch_rt`

The value for `partition_id` is computed as
`{{workspaceID}}-{{murmur3Hash32Index(userID)}}`, meaning:
- The first part comes from the job’s `WorkspaceID`
- The second part is the partition index derived from the job’s `UserID`

**Note:** Existing jobs in the database will have an empty
`partition_id`. This is acceptable, as the column will not be actively
used for some time.

## Linear Ticket

Resolves PIPE-2557
Resolves PIPE-2559

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.