chore: refactor async destination job metadata capturing (#6528)
🔒 Scanned for secrets using gitleaks 8.28.0

# Description

Refactoring the handling of job parameters and related metadata in the
async destination manager. Streamlining the way job metadata is managed
and passed throughout the codebase, improving performance (less
iterations), maintainability and clarity.
Additionally, capturing job metadata for partition id, which is going to
be required for non-blocking workspace partition migrations.

## Linear Ticket

Resolves PIPE-2601

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.