feat: jobsdb partition buffer flush and switchover (#6559)
🔒 Scanned for secrets using gitleaks 8.30.0

# Description

Extending `JobsDBPartitionBuffer` with a full implementation of
`FlushBufferedPartitions` during which:

- Jobs belonging to the partitions being flushed are transferred from
the buffer JobsDB to the primary JobsDB in batches.
- When the retrieved batches no longer hit size limits, or when the move
operation exceeds its timeout, the switchover phase begins:
- Acquire a write lock (distributed, if necessary) to pause any new
writes.
- Refresh the DS list in the buffer JobsDB, if necessary, to ensure that
all remaining data in the buffer will be included in the move.
- Move any remaining jobs from the buffer JobsDB into the primary
JobsDB.
  - Remove the flushed partitions from the buffered-partitions list.
- Release the lock so that pending and new writes for these partitions
resume directly into the primary JobsDB.


## Linear Ticket

Resolves PIPE-2542

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.